Acción de modificar completa

+++ PHP/Eliminar.php
@@ -1,5 +1,14 @@
 <?php
 require "libs/Conexion.php";
+require "libs/Seguridad.php";
+
+$seguridad = new Seguridad();
+
+if ($seguridad->getUsuario() == null) {
+    header('Location: ../index.php');
+    exit;
+}
+
 $id = $_GET["id"];
 $consulta = mysqli_query($conn,"DELETE FROM equipos WHERE ID_Equipo = '$id'");
